chore: Update melos dependency to version 6.1.0
