+++ apps/openci_runner/lib/src/features/build_job/find_job.dart
@@ -1,13 +1,16 @@
 import 'package:dart_firebase_admin/firestore.dart';
+import 'package:openci_models/openci_models.dart';
 import 'package:runner/src/commands/runner_command.dart';
-import 'package:runner/src/features/build_job/models/job_status.dart';
-import 'package:runner/src/models/build_job/build_job.dart';
 
 Future<BuildJob?> findJob() async {
   final firestore = nonNullFirestoreClientSignal.value;
   final qs = await firestore
       .collection('build_jobs')
-      .where('status', WhereFilter.equal, JobStatus.waiting.name)
+      .where(
+        'buildStatus',
+        WhereFilter.equal,
+        OpenCIGitHubChecksStatus.queued.name,
+      )
       .limit(1)
       .get();
