--- apps/openci_runner/lib/src/features/build_job/models/job_status.dart
@@ -1,6 +1,0 @@
-enum JobStatus {
-  waiting,
-  processing,
-  success,
-  failure,
-}
