+++ apps/openci_runner/lib/src/features/build_job/set_job.dart
@@ -1,13 +1,12 @@
+import 'package:openci_models/openci_models.dart';
 import 'package:runner/src/commands/runner_command.dart';
-import 'package:runner/src/features/build_job/models/job_status.dart';
-import 'package:runner/src/models/build_job/build_job.dart';
 
 Future<void> setJobStatus(
   BuildJob job,
-  JobStatus status,
+  OpenCIGitHubChecksStatus status,
 ) async {
   final firestore = nonNullFirestoreClientSignal.value;
   await firestore.collection('build_jobs').doc(job.id).update({
-    'status': status.name,
+    'buildStatus': status.name,
   });
 }
