+++ apps/openci_runner/pubspec.yaml
@@ -16,10 +16,10 @@ dependencies:
   google_play_api_kit: ^0.2.6
   googleapis: ^13.1.0
   googleapis_auth: ^1.6.0
-  http: ^1.2.1
+  http: ^1.2.2
   json_annotation: ^4.8.1
   mason_logger: ^0.2.5
-  openci_models: ^1.0.3
+  openci_models: ^1.1.3
   path: ^1.9.0
   process_run: ^0.13.3
   pub_updater: ^0.3.0
