+++ packages/github_app/pubspec.yaml
@@ -14,15 +14,16 @@ dependencies:
   dart_jsonwebtoken: ^2.14.0
   dartssh2: ^2.8.2
   dotenv: ^4.2.0
-  freezed_annotation: ^2.4.1
+  freezed_annotation: ^2.4.4
   github: ^9.24.0
   google_play_api_kit: ^0.2.6
   googleapis: ^13.1.0
   googleapis_auth: ^1.6.0
   http: ^1.2.1
-  json_annotation: ^4.8.1
+  json_annotation: ^4.9.0
   mason_logger: ^0.2.5
-  openci_models: ^1.0.10
+  openci_models:
+    path: ../openci_models
   process_run: ^0.13.3
   pub_updater: ^0.3.0
   sentry: ^7.18.0
@@ -30,11 +31,11 @@ dependencies:
 
 dev_dependencies:
   analyzer: ^6.4.1
-  build_runner: ^2.4.4
+  build_runner: ^2.4.13
   build_verify: ^3.1.0
   build_version: ^2.1.1
-  freezed: ^2.4.3
-  json_serializable: ^6.7.1
+  freezed: ^2.5.7
+  json_serializable: ^6.8.0
   melos: ^3.4.0
   mocktail: ^1.0.0
   pedantic_mono: ^1.29.0
