+++ packages/openci_models/lib/openci_models.dart
@@ -1,5 +1,6 @@
 export 'src/workflow/model/workflow_model.dart';
-export 'src/job/domain/job_data_v2.dart';
+export 'src/job/domain/build_job.dart';
 export 'src/target_platform.dart';
 export 'src/github/action_type.dart';
 export 'src/env/env_model.dart';
+export 'src/workflow/workflow_model_v2.dart';
