+++ packages/openci_models/lib/src/github/action_type.dart
@@ -1,4 +1,4 @@
-enum ActionType {
+enum OpenCIGitHubActionType {
   opened,
   reopened,
   edited,
