+++ packages/openci_models/lib/src/target_platform.dart
@@ -1,4 +1,4 @@
-enum TargetPlatform {
+enum OpenCITargetPlatform {
   android,
   ios,
 }
