+++ packages/openci_models/pubspec.yaml
@@ -1,6 +1,6 @@
 name: openci_models
 description: OpenCI Models
-version: 1.0.10
+version: 1.1.3
 repository: https://github.com/open-ci-io/openci/tree/develop/packages/openci_models
 
 environment:
