+++ pubspec.yaml
@@ -4,4 +4,4 @@ environment:
   sdk: ">=3.0.0 <4.0.0"
 
 dev_dependencies:
-  melos: ^6.0.0
+  melos: ^6.1.0
